add a fuzz test
This test accepts a JSON description of two structs, and uses reflection to generate those types and validate that goloose behaves the same as marshaling to JSON and unmarshaling to the output type.

So far it just supports strings, ints, float64s, and (potentially nested) structs. In the future it could maybe support slices, maps, and "any"-typed values.

The fuzzer has been very good at finding bugs in the fuzz test, but it hasn't found anything in goloose itself yet!