more graceful handling of recursive structures
We were chasing pointers all the way through until we got a stack overflow. Add a maximum recursion level so we can return an error instead.

encoding/json does this slightly more sophisticatedly, by keeping track of a map of pointers in its state beyond a certain recursion level and erroring out if it's already seen the same pointer in its recursion stack. That'd be nice to do here, but it'd complicate the code quite a bit.